Fix osx build

+++ include/spdlog/details/fmt_helper.h
@@ -42,7 +42,6 @@ inline void append_int(T n, fmt::basic_memory_buffer<char, Buffer_Size> &dest)
     dest.append(i.data(), i.data() + i.size());
 }
 
-
 template<size_t Buffer_Size>
 inline void pad2(int n, fmt::basic_memory_buffer<char, Buffer_Size> &dest)
 {
@@ -60,23 +59,21 @@ inline void pad2(int n, fmt::basic_memory_buffer<char, Buffer_Size> &dest)
         dest.push_back('0');
         dest.push_back(static_cast<char>('0' + n));
     }
-    else    // negatives (unlikely, but just in case, let fmt deal with it)
+    else // negatives (unlikely, but just in case, let fmt deal with it)
     {
         fmt::format_to(dest, "{:02}", n);
     }
 }
 
-
-
 template<typename T, size_t Buffer_Size>
 inline void pad_uint(T n, unsigned int width, fmt::basic_memory_buffer<char, Buffer_Size> &dest)
 {
     static_assert(std::is_unsigned<T>::value, "append_uint must get unsigned T");
-    auto digits = fmt::internal::count_digits(n);
-    if(width > digits)
+    auto digits = fmt::internal::count_digits(static_cast<uint64_t>(n));
+    if (width > digits)
     {
-        const char* zeroes = "0000000000000000000";
-        dest.append(zeroes, zeroes + width-digits);
+        const char *zeroes = "0000000000000000000";
+        dest.append(zeroes, zeroes + width - digits);
     }
     append_int(n, dest);
 }
@@ -86,7 +83,6 @@ inline void pad3(T n, fmt::basic_memory_buffer<char, Buffer_Size> &dest)
 {
     pad_uint(n, 3, dest);
 }
-
 
 template<typename T, size_t Buffer_Size>
 inline void pad6(T n, fmt::basic_memory_buffer<char, Buffer_Size> &dest)
@@ -99,8 +95,6 @@ inline void pad9(T n, fmt::basic_memory_buffer<char, Buffer_Size> &dest)
 {
     pad_uint(n, 9, dest);
 }
-
-
 
 // return fraction of a second of the given time_point.
 // e.g.
